feat(core): ViewRuntime hit grid mouse dispatch, grapheme pool rendering, palette DSL
- Integrate HitGrid into ViewRuntime for z-order aware mouse dispatch with action resolution
- Add WidgetTree::hit_registration_order() for overlay-aware hit area registration
- Migrate command palette from raw buffer drawing to declarative View DSL (overlay/panel/separator)
- Add GraphemePool support to TextLineWidget rendering path
- Export overlay/panel/separator/rich_text/fill/input from prelude
- Fix box drawing inner dimensions calculation
- Clean up if-guard patterns across codebase for clippy compliance
- Add tests for overlay positioning, action dispatch, and grapheme pool

+++ tests/common/pty.rs
@@ -311,7 +311,7 @@ pub fn spawn_pty(config: &PtyConfig) -> io::Result<PtyResult> {
             libc::setsid();
 
             // Set controlling terminal
-            libc::ioctl(slave_fd, libc::TIOCSCTTY, 0);
+            libc::ioctl(slave_fd, libc::TIOCSCTTY as u64, 0);
 
             // Redirect stdin/stdout/stderr to slave
             libc::dup2(slave_fd, 0);
